feat: move path for unfollowing user
BREAKING CHANGE: Path for unfollowing user from `DELETE
/users/:id/unfollow` was moved to `DELETE /buddies/:id`. The reason
behind this changes was to follow the REST-way.

+++ app/controllers/api/v1/buddies_controller.rb
@@ -1,9 +1,10 @@
 class Api::V1::BuddiesController < ApplicationController
-  before_action :authenticate_request, only: [:create, :unfollow]
-  before_action :set_user, only: [:create, :unfollow]
+  before_action :authenticate_request, only: [:create, :destroy]
 
   def create
-    buddy = @auth_user.buddies.build(buddy: @user)
+    user = User.find(params[:user_id])
+
+    buddy = @auth_user.buddies.build(buddy: user)
 
     if buddy.save
       created_response(data: { buddy: buddy })
@@ -12,16 +13,12 @@ class Api::V1::BuddiesController < ApplicationController
     end
   end
 
-  def unfollow
-    buddy = @auth_user.buddies.find_by!(buddy: @user)
+  def destroy
+    buddy = Buddy.find(params[:id])
+    return forbidden_response unless @auth_user.id == buddy.user_id
+
     buddy.destroy
 
     ok_response(data: { buddy: buddy })
-  end
-
-  private
-
-  def set_user
-    @user = User.find(params[:user_id])
   end
 end
